Final code

+++ install/dynamixel_sdk_custom_interfaces/local/lib/python3.10/dist-packages/dynamixel_sdk_custom_interfaces/msg/_set_current_s.c
@@ -65,7 +65,7 @@ bool dynamixel_sdk_custom_interfaces__msg__set_current__convert_from_py(PyObject
       return false;
     }
     assert(PyLong_Check(field));
-    ros_message->current = (int32_t)PyLong_AsLong(field);
+    ros_message->current = (int16_t)PyLong_AsLong(field);
     Py_DECREF(field);
   }
